requirements1.txt
asyncpg==0.30.0
fastapi==0.115.8
Jinja2==3.1.5
pydantic==2.10.6
uvicorn==0.34.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
numpy==2.2.2
aiortc==1.10.1
av==13.1.0
pyyaml==6.0.2